feat: block based concurrently mutable file implementation
This is the new "file" interface for mutable block-addressed files. Blocks are
written into blobs which are applied in "ranges" to the file. Newer writes take
precedence over older writes. Logic is added to jump around and seek to the
correct position in older blobs when writes overlap.

The writer wraps the Handle with the block transaction and allows for writing
and seeking properly.

Still to-do is:

 - compacting logic
 - validation

Signed-off-by: Christian Stewart <christian@paral.in>