refactor(redis): use in-memory cache for consistent tx behavior
Use the txcache mechanism to ensure that reads performed during a write
transaction will be consistent.

Issue revealed by new e2e tests of the volume + fixed.

Example:

 - NewTransaction(write)
 - Get(key) -> not exists
 - Set(key) -> ok
 - Get(key) -> not exists (should exist)

Fixed with this patch. Also, if the connection is lost halfway through the
transaction, there is some attempt to re-try the transaction so far against
a new Redis conn before giving up.

Signed-off-by: Christian Stewart <christian@paral.in>