feat(kvtx): genjidb: implement sql-compatible store on kvtx
Implements a kvtx-backed store that passes the GenjiDB test suite.

GenjiDB implements a sql compatible store on top of a K/V store.

This creates a sql-compatible store on top of kvtx, which can be layered onto
IAVL or Btree algorithms in a p2p block dag as part of a larger structure.

Signed-off-by: Christian Stewart <christian@paral.in>