feat(block): cursor: bucket override, detach, ephemeral
It may be important to release a sub-graph of a block transaction from memory in
certain cases. For example: when iterating through a MySQL database, keeping the
block graph in memory for portions of the graph that are not "dirty" (changed)
will result in keeping effectively the entire database of references in memory.
This is going to be way too much memory usage / churn during these transactions.

In addition to adding some Release call to block cursors which would release the
reference to the block graph, and release all of the unreferenced block graph,
excluding the blocks that were marked as dirty:

New call for the block cursor: Detach() -> *Cursor which returns a "detached"
cursor with no parent located at the block, with a brand new block sub-graph
(not sharing the cached sub-blocks from the previous cursor). The detached
sub-tree

Detach(ephemeral) - if ephemeral, the sub-tree has no associated block
transaction. If !ephemeral, returns a new Block transaction rooted at the
cursor's previous location. The new cursor will behave normally but will not be
associated with the old block transaction, nor share its cache.

In ephemeral mode (cursor.t == nil)

 - NewTransaction -> NewCursor = c1
 - c2 = c1.FollowRef(2, ref)
 - c3 = c2.FollowRef(2, ref)
 - c2 = _
 - // c2 is released from memory
 - // there are no block graph references created

Consequences of this mode:

 - no write operations would work
 - Parent() will always return nil
 - markDirty does nothing
 - Write() does nothing (and returns error)
 - refHandles is not filled in the cursor
 - FollowRef will still be de-duplicated, but its a DAG (not two-way graph)

Purpose of ephemeral mode: some block DAGs, like File, Blob, sql Row, have read
operations which fetch data from the network from thousands of blocks. It's not
necessary to maintain a block graph in memory for these operations, and locking
/ unlocking / updating the block graph is unnecessarily slow. The optimized path
is Follow -> Fetch -> Unmarshal -> Repeat, while forgetting the parent blocks
along the way. By maintaining only parent -> child refs, we can deduplicate the
Follow call while allowing the Go runtime to garbage-collect the parent block
handles accordingly when they are released by the code.

Signed-off-by: Christian Stewart <christian@paral.in>