feat(world): changelog batching, bloom filter
Purposes of changelog:

 - Watchers: list of subscribers of events relating to list of keys
   - fast-forward changelog, check bloom filter for all watchers covered keys,
     if no match skip changelog entry (world change batch).
   - each changelog entry is a series of operations of the same type
 - WaitRev: internally detect when a change occurs to a key
   - current implementation: check (with GetObject) every time world changes
     - optimize: check changelog and detect when to recheck object state

Add batching of multiple WorldChange entries into a block and world seqno. Add
bloom filters to quickly detect if a change impacted a object key. Add a
2-dimensional linked-list for the changelog headers.

The idea is that creating the changelog is a deterministic operation with the
block graph. Traversal of the changelog multiple times to validate is not
necessary. On the other hand, when reading the changelog, you want to quickly
skip any batches that do not affect whatever watchers are active. Traversing the
changelog is necessary anyway to determine which changes were made.

Reason: batch delete objects with a given prefix:

 - If there is 1 entry in the changelog for each object, this is slow
 - Need to batch multiple ChangeLog entries into a single block
 - Instead have 1 entry for the batch deletion
   - KeyFilters: contains checks to see if a key might have been affected
   - ChangeBatch: linked-list of batches of Change
     - (in arbitrary order)

When writing, the block graph can buffer a list of operations in memory, one for
each type of operation. Defer updating the changelog blocks until committing
changes. When committing the operations, batch the changelog in this order:

 1. Object set changes (create, set root ref)
 2. Graph quad create and delete changes.
 3. Object delete changes (delete object).

The object delete changes can only exist in the changelog once DeleteGraphObject
has been applied. So by following the above order of operations, we can always
ensure a transaction has a consistent output, using the above order of ops.

watchers:

 - if change_batch.prev_ref is empty, skip bloom filter checks
 - check bloom filter (use one with 0% false negative)
 - check key prefixes
 - if checks passed: traverse change_batch entries

main changes from current model:

 - WorldChangeLL block (change batching)
 - Bloom filters
 - Batches of WorldChange (multiple into a single block)
 - all of the field numbers changed (Check FollowRef, etc)
 - Change dedupe
   - optimize block graph WorldChange writer:
   - keep list of pending ops, delete ineffectual changes
     - set key K, set key K again (immediately)
     - set key K, delete key K (in same op)
   - Generate change log and store with order of operations

The primary difference from the old implementation is now up to 8 WorldChange
will be directly inlined into the WorldChangeLL (rather than 1 single change). A
mechanism for removing ineffectual world operations is added. And readers will
be able to skip entire batches of WorldChange if they do not meet key prefix,
change type, and bloom filter smoke tests.

+--------------+
| ChangeLL I=5 | --> | Change 5 | --> | Change 4 | --> | Change 3 | --> ...
+--------------+
       |
+--------------+
| ChangeLL I=4 | --> | Change 2 |
+--------------+
       |
+--------------+
| ChangeLL I=3 | --> | Change 1 |
+--------------+

Wwhen layering two World together into a View (which most likely will implement
WorldState), the revision IDs of the objects will be difficult to determine. In
other words: the revision ID being sequential is only a valid assumption to make
during the span of a WorldState handle existing. Revision IDs shouldn't be
stored and depended on to be sequential between executions of the system,
because the underlying storage might have changed and/or rolled back.

Signed-off-by: Christian Stewart <christian@paral.in>