feat: task: resolve and wait for inputs
While the task is Pending, the Task Controller can update the target and inputs
of the task with the UPDATE_INPUTS transaction.

Updating the target and/or inputs returns the Task to PENDING state and cancels
any ongoing Pass.

The transaction diffs the current & updated Inputs to ensure we don't cancel a
Task after a no-op changeset is applied.

While waiting for unset inputs (world objects that don't exist, for example) the
Task controller will progressively update the ValueSet.

Signed-off-by: Christian Stewart <christian@paral.in>