refactor: explicitly expose host volume to plugins
Instead of using the Bifrost RPC invoker for plugin RPCs, register all of the
services the plugin will have access to.

Instead of calling LookupRpcService for the AccessVolumes RPC service every time
the plugin wants to send a volume RPC, look up the volume in advance, and
construct a single ProxyVolume specifically for that volume. This is much less
"chatty" than spamming directives every time the plugin sends a request.

Signed-off-by: Christian Stewart <christian@paral.in>