feat: implement bldr publish command
publish: copy manifests from one remote to another, transforming the blocks
according to the desired transform config. If no transform config is configured,
inherit the transform from the world ObjectRef (via AccessWorldState).

Note that if we change the transform config for the world but keep the same
manifests as before, the manifests will then have the old transform config, and
the world the new one, and newer manifests having the new transform config. So,
this is a way of changing the keys

Alternatively you can delete the world entirely, re-create with a new transform
config, and then run "publish" again to re-encrypt all the manifests.

Goals:

 - Allow configuring the transform config for each set of manifests.
 - Inherit the world transform config if none configured.
 - Make the command line make sense as well.

Current command line flags:

 - remote: remote config to use for storage
   - when publishing is this the source or destination remote?
   - in other contexts "remote" is the one used for build artifacts
   - making this the source remote seems to make the most sense

Design:

 - command line remote is the source world to publish from
 - specify list of publish targets with -p flag
 - each publish target can:
   - override the remote.objectKey for the source object key
   - override the storage for all manifests or per manifest
     - transform config

Signed-off-by: Christian Stewart <christian@aperture.us>