refactor: implement js/wasm build and split devtool and dist entry
Restructure build for new js/wasm approach.

Running with a WebSocket works:

yarn start:web

Loading the web runtime into the browser works:

yarn start:web:wasm

Compiling the dist entrypoint for wasm works but needs work to copy the
index.html and the runtime-wasm.ts files for that platform.

All need the implementation of the web plugin host, which is planned and will be
implemented in the next commit.

Signed-off-by: Christian Stewart <christian@aperture.us>