feat: implement loading WebAssembly plugins
WebAssembly plugins with the devtool now is supported:

yarn start:web:wasm

The browser communicates with the devtool via a WebSocket.

Information is passed to the web devtool entrypoint via a HTTP endpoint.

The devtool builds the manifests on-demand and watches for changes.

The browser hot-reloads plugins when necessary.

The WebAssembly is loaded by a TypeScript shim compiled at build time.

The plugin wasm is served via the ServiceWorker.

The workers communicate with each other by passing MessagePort.

Unified the WebRuntimeClient and WebDocumentTracker types.

Added retry to web document monitoring rpc.

Clarified requirements for passing message port via context bridge.

Signed-off-by: Christian Stewart <christian@aperture.us>