feat: monitor visibility api and suspend web doc if hidden
Watch visibility API. Immediately when not-visible:

- Mark the page as !visible and unregister all WebView with the remote.
- Reset all of the loaded WebViews (set render mode to none & unload UI).
- This will cause most of the WebView to be unregistered.
  - While running in background tab mode we will not inform the WebRuntime.
- The document should still be able to respond to pings.
- If the WebDocument is backgrounded & loses connection, do not re-connect.
- This should all happen within the 1000ms grace period.

When becoming visible again:

- Re-establish connection with the WebRuntime (if connection was lost).
- Mark the page as visible and send the list of web views.
- The Go runtime will re-mount all UI into the WebViews.

This way the document is still mounted and can be used for starting web workers,
but also will not have any content rendered (no WebView), fixing issues with
delayed rpcs under throttling from being a background tab.

Workers are not yet throttled: https://issues.chromium.org/issues/40545662 so
there is no need to worry yet about unloading those when the page is not
visible. Otherwise we would need some logic to suspend the runtime when all of
the pages become not-visible, an exercise left for a future date.

Signed-off-by: Christian Stewart <christian@aperture.us>