refactor: remove unloading web views in hidden document
The problem with unloading WebView in a hidden document is that the contents of
the page is effectively reset every time the user tabs out.

This resets all inputs on the page as well as any scroll state.

We could put the extra work to serialize and restore all of that state, but that
would be a massive pain point in app development with bldr.

The background tabs are just throttled, not fully frozen.

Instead of unmounting the web views, we can keep them mounted and increase the
timeouts to be very long for WebRuntime.

In normal cases the before unload callback will clean up the web document by
sending a close notification.

Signed-off-by: Christian Stewart <christian@aperture.us>