refactor: move plugin assets to /b/pa/
In preparation for TypeScript plugins, the logic for serving the plugin assets
filesystem is complex, and we want it in a common place run by the plugin host
instead of the plugin itself. This way we can have unified behavior serving the
plugin assets filesystem across all the different types of plugin
implementations without needing to include a http filesystem in each.

Signed-off-by: Christian Stewart <christian@aperture.us>