feat: implement wazero-quickjs plugin host and sandbox
Runs js plugins within a quickjs within wazero.

Runs a yamux two-way stream over stdin for Go to Js communciation and /dev/out
within the wasi sandbox for js to go communication.

Successfully tested reading plugin information via RPC call to the plugin host
from the js plugin.

Next step: we need two different platform IDs, currently represented by "web":

- **js**: javascript plugins running in a js sandbox
- **wasm**: wasm plugins running within wasi sandbox
  - Go plugins running within wasm

The current issue is we need a js shim to run the Go wasm plugins.

I think the best solution to this is to ship the web browser targeted js/wasm Go
plugins as native/js/wasm (GOOS=js GOARCH=wasm) and have a separate plugin host
for running these which bundles the js shim for the wasm binary within.

Eventually we can switch to shipping the Go binaries as native/wasip1/wasm which
can be executed within the wazero VM as well.

We need a new platform ID for js-only plugins, which are imported as modules and
executed with `await default(backendAPI)`.

For now, this commit just implements the proof of concept of wazero + quickjs.

Signed-off-by: Christian Stewart <christian@aperture.us>